requirements.txt
setuptools>=68.0.0
wheel>=0.40.0
flask==3.0.2
werkzeug>=3.0.0
numpy>=1.26.0
torch>=2.1.0
transformers>=4.36.0
python-dotenv==1.0.1
pytest==8.0.2
black==23.7.0
flake8==6.1.0
mypy==1.5.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-caching==2.1.0
pandas>=2.2.0
scikit-learn>=1.4.0
gunicorn==21.2.0
python-json-logger==2.0.7
requests==2.31.0
tqdm==4.66.2
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=4.25.1 